Rename HeaderBlock
