+++ src/blocks/HeaderBlock/HeaderBlock.stories.tsx
@@ -6,7 +6,7 @@ import StoryRouter from 'storybook-react-router'
 import { Provider } from 'constate'
 import { globalStoryMock, link, minimalColorMap } from 'utils/storybook'
 import { fallbackLocale } from 'l10n/l10n-utils'
-import { HeroBlock } from '../HeroBlock/HeroBlock'
+import { HeroBlock } from 'blocks/HeroBlock/HeroBlock'
 import { Header } from '.'
 
 export default {
